+++ generator/src/declaration.rs
@@ -81,6 +81,7 @@ pub enum Type {
         kind: Mutability,
         lifetime: Option<Lifetime>,
     },
+    ImplRepr(Box<Type>),
 }
 
 impl Type {
@@ -219,6 +220,10 @@ impl Type {
                     Mutability::Mut => quote! { &#lifetime mut [#of] },
                 }
             }
+            Type::ImplRepr(ty) => {
+                let ty = ty.rust_type(source);
+                quote! { impl crate::Repr<#ty> }
+            }
         }
     }
 
@@ -243,7 +248,7 @@ impl Type {
     pub fn pointer_to_slice(self, lifetime: Option<Lifetime>, source: &Source) -> Type {
         match self {
             Type::Pointer { to, kind } => Type::Slice {
-                of: Box::new(to.builderify(source)),
+                of: Box::new(to.apply_impl_repr(source)),
                 kind,
                 lifetime,
             },
@@ -251,7 +256,7 @@ impl Type {
         }
     }
 
-    pub fn builderify(self, source: &Source) -> Type {
+    pub fn apply_impl_repr(self, source: &Source) -> Type {
         if let Type::Named(Name::Type(name)) = &self {
             let structure = source.find_structure(&name).or_else(|| {
                 source
@@ -261,7 +266,7 @@ impl Type {
 
             if let Some(structure) = structure {
                 if structure.qualifies_as_builder() {
-                    return Type::Named(Name::Type(name.clone().set_builder(true)));
+                    return Type::ImplRepr(Box::new(Type::Named(Name::Type(name.clone()))));
                 }
             }
         }