+++ generator/src/items/structures/builders.rs
@@ -443,6 +443,8 @@ impl Structure {
                     &mut self.0
                 }
             }
+
+            unsafe impl crate::Repr<#inner_ident> for #ident<'_> {}
         }
     }
 }