+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Introduced unsafe `Repr` trait to allow either
+builder or non-builder variants of Vulkan structs to be passed into APIs
 
 ## [0.18.0+174] - 2021-03-29
 ### Added